+++ examples/example.js
@@ -38,8 +38,8 @@ const dbConfig = require('./dbconfig.js');
 // On Windows and macOS, you can specify the directory containing your Oracle
 // Client Libraries.  If this is not done, then a standard search heuristic is
 // used, see the node-oracledb documentation.
-// oracledb.initOracleClient({ libDir: 'C:\instantclient_19_3' });             // Windows
-// oracledb.initOracleClient({ libDir: '/Users/myname/instantclient_19_3' });  // macOS
+// oracledb.initOracleClient({ libDir: 'C:\instantclient_19_3' });                    // Windows
+// oracledb.initOracleClient({ libDir: '/Users/your_username/instantclient_19_3' });  // macOS
 
 async function run() {
   let connection;
@@ -105,9 +105,10 @@ async function run() {
 
     // For a complete list of options see the documentation.
     options = {
-      outFormat: oracledb.OUT_FORMAT_OBJECT   // query result format
-      // extendedMetaData: true,   // get extra metadata
-      // fetchArraySize: 100       // internal buffer allocation size for tuning
+      outFormat: oracledb.OUT_FORMAT_OBJECT,   // query result format
+      // extendedMetaData: true,               // get extra metadata
+      // prefetchRows:     100,                // internal buffer allocation size for tuning
+      // fetchArraySize:   100                 // internal buffer allocation size for tuning
     };
 
     result = await connection.execute(sql, binds, options);