+++ test/prefetchRows.js
@@ -1,4 +1,4 @@
-/* Copyright (c) 2017, 2018, Oracle and/or its affiliates. All rights reserved. */
+/* Copyright (c) 2017, 2020, Oracle and/or its affiliates. All rights reserved. */
 
 /******************************************************************************
  *
@@ -21,123 +21,219 @@
  * NAME
  *   147. prefetchRows.js
  *
- * DESCRIPTION
- *   prefetchRows is deprecated and removed from the oracledb object.
- *   Users can set it but it has no effect since it treated like a user property.
- *
  *****************************************************************************/
 'use strict';
 
-var oracledb = require('oracledb');
-var should   = require('should');
-var dbConfig = require('./dbconfig.js');
+const oracledb = require('oracledb');
+const should   = require('should');
+const assert   = require('assert');
+const dbconfig = require('./dbconfig.js');
 
 describe("147. prefetchRows.js", function() {
 
-  var connection = null;
-  var defaultVal = oracledb.prefetchRows;
+  let conn;
 
-  before(function(done) {
-    oracledb.getConnection(dbConfig, function(err, conn) {
-      should.not.exist(err);
-      connection = conn;
-      done();
-    });
+  const DefaultPrefetchRows = oracledb.prefetchRows;
+  before(async () => {
+    try {
+      conn = await oracledb.getConnection(dbconfig);
+    } catch (error) {
+      should.not.exist(error);
+    }
   });
 
-  after(function(done) {
-    should.strictEqual(defaultVal, undefined);
-    oracledb.prefetchRows = defaultVal;
-    connection.close(function(err) {
-      should.not.exist(err);
-      done();
-    });
+  after(async () => {
+    try {
+      oracledb.prefetchRows = DefaultPrefetchRows;
+
+      await conn.close();
+    } catch (error) {
+      should.not.exist(error);
+    }
   });
 
-  describe("147.1 oracledb.prefetchRows", function() {
+  afterEach(() => {
+    try {
+      oracledb.prefetchRows = DefaultPrefetchRows;
+    } catch (error) {
+      should.not.exist(error);
+    }
+  });
 
-    var testGlobalOption = function(values, cb) {
-      should.doesNotThrow(
+  it('147.1 set oracledb.prefetchRows to be 0', async () => {
+    try {
+      oracledb.prefetchRows = 0;
+      await conn.execute("select 'foobar' from dual");
+    } catch (error) {
+      should.not.exist(error);
+    }
+  }); // 147.1
+
+  it('147.2 Negative - negative value', () => {
+    try {
+      should.throws(
         function() {
-          oracledb.prefetchRows = values;
-        }
+          oracledb.prefetchRows = -9;
+        },
+        'NJS-004: invalid value for property prefetchRows'
       );
-
-      connection.execute(
-        "select 'oracledb.prefetchRows' from dual",
-        function(err, result) {
-          should.not.exist(err);
-          should.strictEqual(result.rows[0][0], "oracledb.prefetchRows");
-          cb();
-        }
-      );
-    };
-
-    it("147.1.1 oracledb.prefetchRows = 0", function(done) {
-      testGlobalOption(0, done);
-    });
-
-    it("147.1.2 oracledb.prefetchRows = NaN", function(done) {
-      testGlobalOption(NaN, done);
-    });
-
-    it("147.1.3 oracledb.prefetchRows = undefined", function(done) {
-      testGlobalOption(undefined, done);
-    });
-
-    it("147.1.4 oracledb.prefetchRows = null", function(done) {
-      testGlobalOption(null, done);
-    });
-
-    it("147.1.5 oracledb.prefetchRows = random string", function(done) {
-      testGlobalOption("random string", done);
-    });
-
-    it("147.1.6 oracledb.prefetchRows = Boolean", function(done) {
-      testGlobalOption(true, done);
-    });
-
+    } catch (error) {
+      should.not.exist(error);
+    }
   });
 
-  describe("147.2 execute() option prefetchRows", function() {
-
-    var testExecOption = function(values, cb) {
-      connection.execute(
-        "select 'prefetchRows' from dual",
-        [],
-        { prefetchRows: values },
-        function(err, result) {
-          should.not.exist(err);
-          should.strictEqual(result.rows[0][0], "prefetchRows");
-          cb();
-        }
+  it('147.3 Negative - NaN', () => {
+    try {
+      should.throws(
+        function() {
+          oracledb.prefetchRows = NaN;
+        },
+        'NJS-004: invalid value for property prefetchRows'
       );
-    };
+    } catch (error) {
+      should.not.exist(error);
+    }
+  });
 
-    it("147.2.2 prefetchRows = 0", function(done) {
-      testExecOption(0, done);
-    });
+  it('147.4 Negative - undefined', () => {
+    try {
+      should.throws(
+        function() {
+          oracledb.prefetchRows = undefined;
+        },
+        'NJS-004: invalid value for property prefetchRows'
+      );
+    } catch (error) {
+      should.not.exist(error);
+    }
+  });
 
-    it("147.2.2 prefetchRows = NaN", function(done) {
-      testExecOption(NaN, done);
-    });
+  it('147.5 Negative - null', () => {
+    try {
+      should.throws(
+        function() {
+          oracledb.prefetchRows = null;
+        },
+        'NJS-004: invalid value for property prefetchRows'
+      );
+    } catch (error) {
+      should.not.exist(error);
+    }
+  });
 
-    it("147.2.3 prefetchRows = undefined", function(done) {
-      testExecOption(undefined, done);
-    });
+  it('147.6 Negative - random string', () => {
+    try {
+      should.throws(
+        function() {
+          oracledb.prefetchRows = "random string";
+        },
+        'NJS-004: invalid value for property prefetchRows'
+      );
+    } catch (error) {
+      should.not.exist(error);
+    }
+  });
 
-    it("147.2.4 prefetchRows = null", function(done) {
-      testExecOption(null, done);
-    });
+  it('147.7 Negative - Boolean', () => {
+    try {
+      should.throws(
+        function() {
+          oracledb.prefetchRows = true;
+        },
+        'NJS-004: invalid value for property prefetchRows'
+      );
+    } catch (error) {
+      should.not.exist(error);
+    }
+  });
 
-    it("147.2.5 prefetchRows = random string", function(done) {
-      testExecOption("random string", done);
-    });
+  it('147.8 execute() option, value of 0', async () => {
+    try {
+      const options = { prefetchRows: 0 };
+      await conn.execute("select 'prefetchRows' from dual", [], options);
+    } catch (error) {
+      should.not.exist(error);
+    }
+  });
 
-    it("147.2.6 prefetchRows = Boolean", function(done) {
-      testExecOption(true, done);
-    });
+  it('147.9 Negative - negative value', async () => {
+    try {
+      const options = { prefetchRows: -10 };
+      await assert.rejects(
+        async () => {
+          await conn.execute("select 'prefetchRows' from dual", [], options);
+        },
+        /NJS-007/
+      );
+      // NJS-007: invalid value for "prefetchRows" in parameter 3
+    } catch (error) {
+      should.not.exist(error);
+    }
+  });
 
+  it('147.10 Negative - NaN', async () => {
+    try {
+      const options = { prefetchRows: NaN };
+      await assert.rejects(
+        async () => {
+          await conn.execute("select 'prefetchRows' from dual", [], options);
+        },
+        /NJS-007/
+      );
+    } catch (error) {
+      should.not.exist(error);
+    }
+  });
+
+  it('147.11 execute() option, undefined, get overrided by global attribute', async () => {
+    try {
+      const options = { prefetchRows: undefined };
+      await conn.execute("select 'prefetchRows' from dual", [], options);
+    } catch (error) {
+      should.not.exist(error);
+    }
+  });
+
+  it('147.12 Negative - null', async () => {
+    try {
+      const options = { prefetchRows: null };
+      await assert.rejects(
+        async () => {
+          await conn.execute("select 'prefetchRows' from dual", [], options);
+        },
+        /NJS-007/
+      );
+    } catch (error) {
+      should.not.exist(error);
+    }
+  });
+
+  it('147.13 Negative - random string', async () => {
+    try {
+      const options = { prefetchRows: 'random string' };
+      await assert.rejects(
+        async () => {
+          await conn.execute("select 'prefetchRows' from dual", [], options);
+        },
+        /NJS-007/
+      );
+    } catch (error) {
+      should.not.exist(error);
+    }
+  });
+
+  it('147.14 Negative - Boolean', async () => {
+    try {
+      const options = { prefetchRows: true };
+      await assert.rejects(
+        async () => {
+          await conn.execute("select 'prefetchRows' from dual", [], options);
+        },
+        /NJS-007/
+      );
+    } catch (error) {
+      should.not.exist(error);
+    }
   });
 
 });